requirements.txt
requests
pandas
streamlit
python-dotenv
tenacity
plotly
groq
